New create() method to mirror node implementation
Spent about a half hour trying to figure out how to send a picture.  The current implementation of passing a MediaFile was kind of confusing, especially since this only using the content anyway.

Currently, just to send a simple picture message, I have to add these 5 lines to create a MediaFile:
 
        String mediaUrl =  "https://s3.amazonaws.com/bwdemos/logo.png";
        Map<String,String> media = new HashMap<String, String>();
        media.put("content",mediaUrl);
        JSONObject json = new JSONObject(media);
        MediaFile file = new MediaFile(client, json);

Would be a nice QOL improvement to be able to just call it with the URL like you can with the node library.